requirements.txt
Pillow
bs4
cowpy
emoji
gitpython
googletrans
gtts
heroku3
humanize
lyricsgenius
psycopg2-binary
pybase64
pyrogram
python-dotenv
requests
selenium
sqlalchemy
tgcrypto
urbandict
wikipedia
